use base classes instead of mixins
